Error Handling

+++ server.js
@@ -9,15 +9,11 @@ const bodyParser = require('body-parser');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 
-//FRONTEND RUNNER
-const frontendPath = path.join(__dirname, "frontend");
-app.use(express.static(frontendPath));
-
 // CONSTANTS //
 const bashPath = 'C:/Program Files/Git/bin/bash.exe';
 const unzipscriptPath = 'shell-scripts/tar-unzip.sh';
 const tempFolderPath = 'tempFolder';
-const deletescriptPath = path.join(__dirname, '/shell-scripts/delete-unzipped.sh');
+const deletescriptPath = 'shell-scripts/delete-unzipped.sh';
 
 //FUNCTIONS
 async function unzipTarFunction(...args){
@@ -106,34 +102,52 @@ async function deleteUnzipped(...deleteArgs){
   })
 }
 
+async function problemIDExists(directoryPath) {
+  try {
+    await fs.access(directoryPath, fs.constants.F_OK);
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
 //APIs
 app.post("/getDirectoryDetailsAll", async (req, res) => {
   try {
     const { projectID, projectName } = req.body;
-    const directoryPath = path.join('efs-mount-point/workspaces', projectID, projectName+'.tar');
+    let directoryPath = path.join('efs-mount-point/workspaces', projectID);
+    const problemIDExistsRes = await problemIDExists(directoryPath);
 
-    // Await unzip process completion
-    const args = [directoryPath, tempFolderPath];
-    const unzipRes = await unzipTarFunction(...args);
-    console.log(JSON.stringify(unzipRes.toString()) + '\n');
-    
-    // Get directory details
-    const directoryPathToGetDetails = path.join("tempFolder", projectName);
-    const directoryDetails = await getDirectoryDetails(directoryPathToGetDetails, projectName);
+    if(!problemIDExistsRes){
+      res.status(404).json({ message: "Problem ID Directory Doesn't Exist." });
+    } else{
+      directoryPath = path.join(directoryPath, projectName+'.tar');
+      // Await unzip process completion
+      const args = [directoryPath, tempFolderPath];
+      const unzipRes = await unzipTarFunction(...args);
+      console.log(JSON.stringify(unzipRes.toString()) + '\n');
+      
+      // Get directory details
+      const directoryPathToGetDetails = path.join("tempFolder", projectName);
+      const directoryDetails = await getDirectoryDetails(directoryPathToGetDetails, projectName);
 
-    // Delete Temp Folder
-    const deleteArgs = [tempFolderPath, projectName];
-    const deleteUnzippedRes = await deleteUnzipped(...deleteArgs);
-    console.log(JSON.stringify(deleteUnzippedRes.toString()));
+      // Delete Temp Folder
+      const deleteArgs = [tempFolderPath, projectName];
+      const deleteUnzippedRes = await deleteUnzipped(...deleteArgs);
+      console.log(JSON.stringify(deleteUnzippedRes.toString()));
 
-    res.send(directoryDetails);
+      res.send(directoryDetails);
+    }
   } catch (error) {
     console.error(error)
     res.status(500).send(`Internal server error: ${error}`);
   }
 });
 
-//Frontend
+//FRONTEND RUNNER
+const frontendPath = path.join(__dirname, "frontend");
+app.use(express.static(frontendPath));
+
 app.get("/", (req, res) => {
   app.use(express.static(frontendPath));
   res.send("success");
@@ -145,7 +159,6 @@ problemID and projectName form Frontend.
   Done - 1. Frontend should be sending problem and project id/name in req body. (get req - req body won't work - only query or path).
     
   Pending - 2. handle error 
-                maybe problem id is not there.
-                maybe Project name .js .cpp doesn't exist.
-  
+    maybe problem id is not there.
+    maybe Project name .js .cpp doesn't exist.
 */
